+++ vimiv/api/settings.py
@@ -475,6 +475,22 @@ class thumbnail:  # pylint: disable=invalid-name
         True,
         desc="Save new thumbnails to disk in the shared icon cache for later use",
     )
+    load_behind = IntSetting(
+        "thumbnail.load_behind",
+        -1,
+        desc="Max thumbnails to render behind the selected one. -1 is no limit",
+    )
+    load_ahead = IntSetting(
+        "thumbnail.load_ahead",
+        -1,
+        desc="Max thumbnails to render ahead of the selected one. -1 is no limit",
+    )
+    unload_threshold = IntSetting(
+        "thumbnail.unload_threshold",
+        -1,
+        desc="Don't unload currently rendered thumbnails unless the number loaded is"
+        " above this threshold. -1 is no limit",
+    )
 
 
 class slideshow:  # pylint: disable=invalid-name